requirements.txt
blessings==1.6
bpython==0.15
curtsies==0.2.6
Django==1.9.3
djangorestframework==3.3.2
docker-py==1.7.2
greenlet==0.4.9
Pygments==2.1.3
requests==2.9.1
six==1.10.0
wcwidth==0.1.6
websocket-client==0.35.0
wheel==0.24.0
django-http-proxy==0.4.3
